Switched ImGui branch to Docking one and Improved (although linking error)
ImGui version of our engine has been switched to the one with docking, and the ImGui layer has been improved, although it still gives a linking error because ImGui compiles as a static library and the engine as a dynamic one (easy fix).

Also, deleted the files we did for ImGuiOpenGLRenderer and fixed a minor issue in LayerStack.h, as well as added a UIRender() function to layers so they can make the UI there.