Implemented DeltaTime and moved stuff to the Editor section
There was a lot of testing/temporary stuff in the EngineApplication, so it was moved into the EditorApplication, which is a bit proper and moved the declaration to use the dedicated graphics card into the Engine's entry point (was in the EditorApp, so wrong!)

Also, implemented DeltaTime for Engine Updates so we can have a unifrom refresh rate independent from the clock speed. However, is pretty basic still, so it should be improved on time.