Couple of errors fixed on Instrumentor & Improved BIND() macro
Instrumentor.h had a couple of problems, the first was when adding different timers to a single scope (the __LINE__ macro was not expanding) and the other was a memory leak which malformed the resulting json because the footer was not called when exiting the app without "ending" the instrumentor. They are now fixed.

Also, it seems that using std::bind() is a bad practice in modern C++, so following a commit and merge from The Cherno's Hazel Engine (#277), this is fixed by using tools of modern c++ (decltype(auto) and std::forward() to forward generic arguments)