Fixed bug on camera deserialization & init engine with CubeScene
There was a bug upon camera deserialization, specially with Projection ones. The thing was that AspectRatio was "nan" because on the first ViewportResize() height (divisor) was 0, and this caused a crash on the first RecalculateMatrix(). Now this is fixed.

Also, commented SaveScene() and the engine will now begin with a loaded scene