Texture Coordinates as VArray defined (dummy commit, just see desc.)
Up until now, the TCoords were defined already in the VArray, but they were hardcoded in the renderer draw function since we were only drawing quads, unlike position, which is defined as "RendererData".

Now UVs are like positions, so it's not "very hardcoded", although in a future they will have to belong to a mesh or similar.

Nevertheless, this is the 1st step to make the UVs node parameter of the material node