FINALLY WORKING: 3D BATCH RENDERER WITH MODELS LOAD
OOOOOOOOOOOH YESSSSSSSSSSSSSSSSSSSSSS, JESUS TAKE THE WHEEL!

This issue I've been with the last 2 weeks is FINALLY FIXEEEEEEEEEED!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

Who wants a Eurocup eh Italy? I have this masterpiece!

So this batched 3D renderer is finally working, necessary for the material editor to work with 3D opjects. And of course, the 3d models load works perfecly :)

Still, there are some improvements to make, but it works :)