Moved Renderer/Resources Serialization + Materials as umap instead of vecs
Just made the materials to be stored as unordered maps instead of vectors, I prefer it that way. Also made some changes to the renderer, now we store "RendererData", not "SceneData", is way a better name

Also moved the renderer and resources serialization, the scene serialization was not the place to do that. They are now in the application